expug v0.0.3 Expug.Transformer

Transforms a node after compilation.

Functions

clause_after(arg1)

Finds out what clauses can follow a given clause.

iex> Expug.Transformer.clause_after("if")
["else"]

iex> Expug.Transformer.clause_after("try")
["catch", "rescue", "after"]

iex> Expug.Transformer.clause_after("cond")
[] # nothing can follow cond

close_clause(children, next)

Closes all a given next clause in the given children.

close_clauses(children)

Closes all possible clauses in the given children.

prelude(arg1)

Get the prelude of a given node

iex> Expug.Transformer.prelude(%{value: "if foo"})
"if"

iex> Expug.Transformer.prelude(%{value: "case derp"})
"case"

iex> Expug.Transformer.prelude(%{value: "1 + 2"})
nil

transform(node)

Transforms a node.