expug v0.0.1 Expug.TokenizerTools

For tokenizers.

def tokenizer(source)
  {[], source, 0}
  |> document()
  |> finalize()
end

def document(state)
  state
  |> eat(~r/^doctype /, :doctype)  # create a token
end

The state

The state begins as a tuple of {[], source, 0}. This is the list of tokens, the source text, and the cursor position (aka, {doc, source, pos}).

eat(state, ~r/^"/, :open_quote)

The eat() function tries to find the given regexp from the source at position pos. If it matches, it returns a new state: a new token is added (:open_quote in this case), and the position pos is advanced.

If it fails to match, it’ll throw a {:parse_error, pos, [:open_quote]}. Roughly this translates to “parse error in position pos, expected to find :open_quote”.

Mixing and matching

Normally you’d make functions for most token types:

def doctype(state)
  state
  |> eat(%r/^doctype/, :doctype, nil)
  |> whitespace()
  |> eat(%r/^[a-z0-9]+/, :doctype_value)
end

You can then compose these functions using:

state
|> one_of([ &doctype/1, &foobar/1 ])
|> optional(&doctype/1)
|> many_of(&doctype/1)

Functions

convert_position(pos, offsets)

Converts a position number n to a tuple {line, col}.

convert_positions(doc, source)

Converts numeric positions into {line, col} tuples.

iex> source = "div\n  body"
iex> doc = [
...>   { 0, :indent, "" },
...>   { 0, :element_name, "div" },
...>   { 4, :indent, "  " },
...>   { 6, :element_name, "body" }
...> ]
iex> Expug.TokenizerTools.convert_positions(doc, source)
[
  { {1, 1}, :indent, "" },
  { {1, 1}, :element_name, "div" },
  { {2, 1}, :indent, "  " },
  { {2, 3}, :element_name, "body" }
]

eat(state, expr)

Eats a token.

state - assumed to be {doc, source, pos} (given by run_tokenizer/2).

expr - regexp expression.

token_name (atom, optional) - token name.

reducer (function, optional) - a function.

Returns { doc, source, pos } too, where doc is transformed via fun.

eat state, ~r/.../, :document
eat state, ~r/.../, :document, nil  # discard it
eat state, ~r/.../, :document, &[{&3, :document, &2} | &1]

# &1 == current state
# &2 == matched String
# &3 == position

eat(state, expr, token_name)

eat(state, expr, token_name, fun)

eat_string(state, expr)

Like eat(), but instead of creating a token, it appends to the last token. Useful alongside start_empty().

state
|> start_empty(:quoted_string)
|> eat_string(~r/^"/)
|> eat_string(~r/[^"]+/)
|> eat_string(~r/^"/)

finalize(arg)

Turns a state tuple ({doc, source, position}) into a final result. Returns either {:ok, doc} or {:parse_error, %{type, position, expected}}. Guards against unexpected end-of-file.

get_parse_errors(arg1)

Extracts the last parse errors that happened.

In case of failure, run_tokenizer() will check the last parse errors that happened. Returns a list of atoms of the expected tokens.

many_of(state, head)

Checks many of a certain token.

many_of(arg, head, tail)

match(expr, remainder)

one_of(state, funs, expected \\ [])

Finds any one of the given token-eater functions.

state |> one_of([ &brackets/1, &braces/1, &parens/1 ])

optional(state, fun)

An optional argument.

state |> optional(&text/1)

optional_many_of(state, head)

scrub_parse_errors(doc)

Gets rid of the :parse_error hints in the document.

start_empty(arg, token_name)

Creates an token with a given token_name. This is functionally the same as |> eat(~r//, :token_name), but using start_empty() can make your code more readable.

state
|> start_empty(:quoted_string)
|> eat_string(~r/^"/)
|> eat_string(~r/[^"]+/)
|> eat_string(~r/^"/)